requirements.txt
numpy
pillow
pycuda
tensorflow[and-cuda]
tf_keras